pyproject.toml
[tool.poetry]
name = "covid-factors"
version = "0.1.0"
description = "COVID Factors"
authors = ["pezon <pezon@uchicago.edu>"]

[tool.poetry.dependencies]
python = ">=3.7.1,<4.0"
requests = "^2.25.1"
pandas = "^1.2.4"
gcsfs = "^2021.4.0"
python-dotenv = "^0.17.1"
click = "^8.0.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
